--- lib/install/install.go
@@ -1,1 +1,0 @@
-package install
