+++ pmgo/config.toml
@@ -1,0 +1,8 @@
+SysFolder = ""
+PidFile = ""
+OutFile = ""
+ErrFile = ""
+
+[Watcher]
+
+[Procs]
