+++ pmgo/lib/install/install.go
@@ -1,0 +1,1 @@
+package install
